ml-service/requirements.txt
fastapi
uvicorn
transformers
torch
scikit-learn
asyncpg
sqlalchemy
python-multipart
python-jose[cryptography]
passlib[bcrypt]
aiofiles
pydantic[email]